* master: (25 commits)
  use lichess.once in ui/voice instead of storedBooleanProp
  don't advertise modals from other modals
  update opening explorer api
  update lowest rating group of fishnet opening book
  update label of lowest /opening rating group
  refactor pubsubhub handler
  better handle at:deleted-entry from pubsubhubb - for #12713
  Fix gap on homepage (#12850)
  form3.hiddenFalse
  further increase explorer cache useful threshold
  increase cache useful threshold
  scala 3.3.0-RC6
  Group imports
  Use new Score
  Fix writing swiss conditions to mongodb when empty
  remove round.id from puzzle activity stream
  code golf for hidden inputs allowing to disable rated
  fix comment for hack about swiss tournaments `conditions.playYourGames`
  API: Allow to retrieve pending simuls when authenticated
  fix slist min width without a wrapper - closes #12795
  ...

+++ modules/opening/src/main/package.scala
@@ -4,5 +4,5 @@ export lila.Lila.{ *, given }
 
 private val logger = lila.log("opening")
 
-type PopularityHistoryAbsolute = List[Int]
+type PopularityHistoryAbsolute = List[Long]
 type PopularityHistoryPercent  = List[Float]
